[Task 3] Duplicate by UUID
When we use a different `column` for the **Route Model Binding**, We have to explicitly mention it in the Route or we have to change the **Route Binding Key** in the model `getRouteKeyName` method override.

We can't change the **Route key Column** in the model in our case, since other Routes rely on `id` being the Route Key. So We have changed the Route definition.